fix: extract MobileFloorNav to get FactoryFloor under 500 lines S1 gate
